fix: Remove @Builder.Default as it adds obscure bugs
@Builder.Default removes the initializer expression and only
adds it in the builder. This is exactly opposite of what we
wanted. We wanted no warnings for lombok and default value
in no constructor instance and no value in builder.
This is a serious misbehaviour and will causes several
`NullPointerException`s in the code, so we can live with a
few lombok warnings